- Internal type for Player::score() was still short, resulting in invalid values presented for players with really high score (for example in SRB2).

+++ src/core/serverapi/player.cpp
@@ -27,7 +27,7 @@ DClass<Player>
 {
 	public:
 		QString name;
-		short score;
+		long score;
 		unsigned short ping;
 		bool spectator;
 		bool bot;
